requirements.txt
atlassian-python-api
langchain-text-splitters
python-dotenv
